회원 가입, 로그인/로그아웃 기능 추가

+++ src/components/board/ReadBoardComponent.jsx
@@ -1,8 +1,9 @@
 import React, { useEffect, useState } from 'react';
-import BoardService from '../services/BoardService';
+import BoardService from '../../services/BoardService';
 import { useParams } from 'react-router-dom';
 import { useNavigate } from "react-router-dom";
 import dayjs from 'dayjs';
+import * as common from "../../utils/common_function";
 
 function ReadBoardComponent() {
     const [board, setBoard] = useState({});
@@ -14,10 +15,10 @@ function ReadBoardComponent() {
         BoardService.getBoardDetail(no).then(res => {
             let response = res.data;
             if (response.resultCode === "SUCCESS") {
-                const data = response.data;
+                const data = JSON.parse(response.data);
                 setBoard(data);
             } else {
-
+                alert(response.resultMessage);
             }
         });
     }, []);
@@ -25,9 +26,9 @@ function ReadBoardComponent() {
     const returnBoardType = (type) => {
         let strType = "";
 
-        if (type == 1) {
+        if (type === 1) {
             strType = "자유게시판";
-        } else if (type == 2) {
+        } else if (type === 2) {
             strType = "QnA 게시판";
         } else {
             strType = "기타";
@@ -59,7 +60,14 @@ function ReadBoardComponent() {
     /**
      * 삭제처리
      */
-    const deleteBoard = () => {
+    const deleteBoard = async () => {
+        if (!await common.checkLogin()) {
+            alert("장시간 입력이 없어 정보를 가져올 수 없습니다. 다시 진행해주세요.");
+            sessionStorage.setItem("IS_LOGIN", "N");
+            navigate("/");
+            return false;
+        }
+
         if (window.confirm("정말 삭제하시겠습니까?")) {
             BoardService.deleteBoard(no).then(res => {
                 let response = res.data;
@@ -67,7 +75,11 @@ function ReadBoardComponent() {
                     alert("삭제되었습니다.");
                     navigate("/board");
                 } else {
-
+                    alert(response.resultMessage);
+                    if (response.apiResultCode === "0099") {
+                        sessionStorage.setItem("IS_LOGIN", "N");
+                        navigate("/");
+                    }
                 }
             });
         }
@@ -86,12 +98,16 @@ function ReadBoardComponent() {
                     </div>
 
                     <div className='row'>
-                        <label> MemberNo </label> : {board.memberNo}
+                        <label> 작성자 </label> : {board.memberId}
                     </div>
                     {returnDate(board.createdTime, board.updatedTime)}
-                    <button className='btn btn-primary' onClick={goUpdateBoard} style={{ marginLeft: "10px" }}>수정하러 가기</button>
                     <button className='btn btn-primary' onClick={goBoardList} style={{ marginLeft: "10px" }}>글 목록으로 이동</button>
-                    <button className='btn btn-primary' onClick={deleteBoard} style={{ marginLeft: "10px" }}>삭제하기</button>
+                    {board.updateAvalYn === "Y" &&
+                        <>
+                            <button className='btn btn-primary' onClick={goUpdateBoard} style={{ marginLeft: "10px" }}>수정하러 가기</button>
+                            <button className='btn btn-primary' onClick={deleteBoard} style={{ marginLeft: "10px" }}>삭제하기</button>
+                        </>
+                    }
                 </div>
             </div>
         </div>
